Fix sel_ldr build on x86-64 on Mac
Fix a type mismatch introduced by r4441 when NaClTlsGetIdx() was
declared in a header.  This was not caught because x86-64 Mac is not
tested on the trybots/buildbots, but apparently the SDK build now
covers it.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1485
TEST=run_hello_world_test (x86-64 Mac not generally supported, so
  run_signal_handler_test does not fully pass there)

Review URL: http://codereview.chromium.org/6612029

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4458 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 5a3126d1fff40802561b8453957f927c2bc1569d

+++ src/trusted/service_runtime/arch/x86_64/nacl_tls_64.c
@@ -222,8 +222,8 @@ void NaClTlsSetIdx(uint32_t tls_idx) {
  * this for x86-32 on OSX, since that's handled by NaCl "Classic"
  * where %gs gets swapped, and we use %gs >> 3 in the asm code.)
  */
-int NaClTlsGetIdx(void) {
-  return (int) (intptr_t) pthread_getspecific(nacl_thread_info_key);
+uint32_t NaClTlsGetIdx(void) {
+  return (intptr_t) pthread_getspecific(nacl_thread_info_key);
 }
